requirements.txt
asgiref==3.7.2
Django==4.2.7
djangorestframework==3.14.0
gunicorn==21.2.0
mysqlclient==2.2.0
packaging==23.2
Pillow==10.1.0
sqlparse==0.4.4
tzdata==2023.3
